Variogram fit adjusting for positional error

This function fit the parameters of a variogram through N-weighetd least squares taking also into account the presence of positional error.

varioadj(data, locations, displacement = "gaussian", delta, kappa, ini, ...)

Arguments

data
A numeric vector of spatial data.

locations
A two column matrix containing the locations (coordinates x and y).

displacement
The type of geomasking to be applied: either "gaussian" or "uniform".

delta
A number that specify the standard deviation of the positional error in the case of Gaussian geomasking or the maximum displacement distance in the case of Uniform geomasking.

kappa
Numerical value for the additional smoothness parameter of the matern correlation function.

ini
Initial values for the parameters to be passed to the optimisation algorithm.

...
Control argumenets for the optimiser.

Value

A numeric vector containg the estimated parameters.